Extract NoContentParser to top level class
